pyproject.toml
[tool.poetry]
name = "graffiti-wall"
version = "0.1.0"
description = "A graffiti wall for the internet"
authors = ["Floozutter <floozutter@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1"
flask-socketio = "^4.2"
Pillow = "^6.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
